==> 01_nanoSIMS_data_reduction.Rmd <==
---
title: "01_CD2F_Data_Reduction"
author: "Tristan Caro"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

## Clear the environment

```{r}
rm(list=ls())
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Introduction

Here we recalculate isotope ratio by ion counts (despite that our processing script already did it).

$$
^2R = \frac{^2H}{^1H}
$$

We also calculate fractional abundance:

$$
^2F = \frac{^2H}{^2H + ^1H}
$$

fractional abundance is, later on, reported as atom percent:

$$
^2F \% = {^2F} * 100 \%
$$

Here we also apply the dead time correction (DTC) to our ion counts. DT is the time between ion pulses in which an electron multiplier is unable to capture additional signal while the previous signal is being processed. When a secondary ion hits the conversion dynode, it ejects an electron that ejects further electrons across the subsequent dynodes. This results in an electron pulse at the pre-amplifier electronics. The time it takes to record this pulse, the detector cannot measure incoming ions arriving at the conversion dynode. This is problematic when the secondary ion count rate is high -- effectively depressing the recorded signal. Our nanoSIMS instrument used the default value of 44 ns DT. The effect of DT should be minimal on our samples and a simple linear correction can be used. This correction is defined as follows where $C_{DTC}$ is ion counts per second (cps) DTC-corrected; $C_{raw}$ is raw cps as reported by the instrument, and $\tau$ is the detector DT in ns. (Adapted from Kilburn, M.R., Clode, P.L. (2014). Elemental and Isotopic Imaging of Biological Samples Using NanoSIMS. In: Kuo, J. (eds) Electron Microscopy. Methods in Molecular Biology, vol 1117. Humana Press, Totowa, NJ. https://doi.org/10.1007/978-1-62703-776-1_33).

$$
C_{DTC} = \frac{C_{raw}}{(1-\tau \cdot C_{raw})}
$$

Alternatively defined as:

$$
N = \frac{N_m}{1 - N_m \cdot \tau / T}
$$

where N is the real number of secondary ions, $N_m$ is the number of secondary ions **m**easured, $\tau$ is deadtime (seconds), and T is dwell time (seconds per pixel).




DTC is applied as follows:

```
# counts of the ion per second
cps_ion = N/(analysis.px.s * PIXELSi)

# dead time corrected (NOTE: ideally px by px but doing it by ROI is sufficient approximation here - max error < 1 ppm)
N_dc = N / (1 - cps_ion * deadtime.ns * 1e-9) 
```


# Setup

## Clear the environment

```{r}
rm(list=ls())
```

## Load Libraries

```{r}
library(tidyverse)
library(ggdist)
source("source/correct_for_deadtime.R")
```

## Load nanoSIMS ROI data from .tsv
```{r}
if (file.exists("cache/roi_data.RDS")) {
  roi_data <- readRDS("cache/roi_data.RDS")
} else {
  tsv_file_list = list.files(path = "data/nanoSIMS/tsv_output/", pattern="*.tsv")
  
  roi_data <- tibble()
  roi_data <- roi_data[-1,] # Remove the garbage row of the tibble
  i = 0
  
  for (tsv in tsv_file_list) {
    temp <- read_tsv(paste0("data/nanoSIMS/tsv_output/", tsv),
                     col_types = cols())
    temp <- temp %>% mutate(filename = tsv)
    roi_data <- bind_rows(roi_data, temp)
    i <- i + 1
    print(paste("File", i, "of", length(tsv_file_list), "loaded."))
  }
  saveRDS(roi_data, "cache/roi_data.RDS")
}

```

## Load cell metadata

```{r}
cell_metadata <- readxl::read_excel("data/other/cell_metadata.xlsx") %>% 
  # Filter out bad nanoSIMS data (ignore == True) and cells that do not have nanoSIMS roi associated with them
  filter(
    ignore != TRUE,
    !is.na(nanoSIMS_file_id)
    ) %>% 
  # Tidy up
  select(-c(description, ignore, raman_spectrum_id)) 
```

## Load nanoSIMS file metadata

```{r}
nanoSIMS_file_metadata <- readxl::read_excel("data/other/cell_metadata.xlsx", sheet = 2)

# generate a smaller df that is just counts/px. This is used for dead time correction
# Counts/px is the pixel dwell time in microseconds (µs)

dwell_time_data <- nanoSIMS_file_metadata %>% 
  select(nanoSIMS_file_id, `Counts/px`) %>% 
  rename(nanoSIMS_acq = nanoSIMS_file_id)
```

## Add organism names, parse nanoSIMS_file_id, add crude metadata

```{r}
roi_data2 <- roi_data %>% 
  # Remove green-marked (not id'd) cells for co-registered dataset
  filter(Group == "red") %>%
  # Add some quick organism id
  mutate(
    organism = case_when(
      str_detect(filename, "Thy") ~ "T. hydrogeniphilus",
      str_detect(filename, "Mb") ~ "M. NSHQ04"
      ),
    # some quick label strength
    label_at2H = case_when(
      organism == "T. hydrogeniphilus" ~ as.numeric(str_sub(filename, start = 5, end = 6)),
      organism == "M. NSHQ04" ~ as.numeric(str_sub(filename, start = 4, end = 5)),
    )
  ) %>% 
  # Parse the nanoSIMS_file_id
  mutate(
    nanoSIMS_file_id = str_sub(filename, end = -21), # this one includes the frame number (e.g. "f0")
    nanoSIMS_acq = str_sub(filename, end = -24) # this one is just the acquisition id (no frame number)
  ) %>% 
  # match the metadata column name
  rename(nanoSIMS_roi_number = ROI)


# Sanity check to make sure all the files are there:
cell_files <- cell_metadata %>% select(nanoSIMS_file_id) %>% unique()
roi_files <- roi_data2 %>% select(nanoSIMS_file_id) %>% unique()


paste("These are the nanoSIMS files that contain ROIs but have zero co-registered cells:", setdiff(roi_files, cell_files))

paste("These are the nanoSIMS_file_ids that exist in the cell_metadata.xlsx but are not accounted for in the .tsvs that were imported:", setdiff(cell_files, roi_files))
  
```


## Join cell_metadata to roi_data


```{r}
detector_DT_ns = 44 # detector deadtime is 44ns, default for nanoSIMS instruments

roi_data_joined <- cell_metadata %>% 
  left_join(
    roi_data2,
    by = c("nanoSIMS_file_id", "nanoSIMS_roi_number")
    # Add in the counts/px data:
  ) %>% left_join(
    dwell_time_data, 
    by = "nanoSIMS_acq"
  ) %>% 
  # Correct for dead time, recalculate 2R
  mutate(
    `2H_dtc` = correct_for_deadtime(
      N = `2H`,
      dwell_time_us = `Counts/px`,
      npx = area,
      deadtime_ns = detector_DT_ns
    ),
    `1H_dtc` = correct_for_deadtime(
      N = `1H`,
      dwell_time_us = `Counts/px`,
      npx = area,
      deadtime_ns = detector_DT_ns
    ),
  ) %>% 
  # Recalculate 2R and calculate 2F
  mutate(
    Ratio_2Hx1H = `2H`/`1H`, # Replace the original 'Ratio_2Hx1H' because that is calculated as 2F in the processing script
    Ratio_2Hx1H_dtc = `2H_dtc`/`1H_dtc`,
    frac_2H = `2H`/ (`1H` + `2H`),
    frac_2H_dtc = `2H_dtc`/ (`1H_dtc` + `2H_dtc`),
    at2H_percent_dtc = frac_2H_dtc * 100,
    at2H_percent = frac_2H * 100
  )

```

## Evaluate DTC data

```{r}
evaluate_dtc <- roi_data_joined %>% 
  mutate(
    diff_2H = `2H_dtc` - `2H`,
    diff_1H = `1H_dtc` - `1H`,
    diff_2R = `Ratio_2Hx1H_dtc` - `Ratio_2Hx1H`,
    diff_2F = frac_2H_dtc - frac_2H,
    diff_dan_2F = `Ratio_2Hx1H` - frac_2H
) %>% 
  select(c(diff_2H, diff_1H, diff_2R, diff_2F, diff_dan_2F))
```


# Save to cache
```{r}
# Cache it!
saveRDS(roi_data_joined, "cache/roi_data_joined.RDS")
```



# OPTIONAL: Explore with some plots

```{r}
p_roi_data <- roi_data %>% 
  filter(organism == "T. hydrogeniphilus") %>% 
  ggplot(
    aes(
      x = label_at2H,
      y = Ratio_2Hx1H,
      color = organism,
      group = label_at2H
    )
  ) +
  geom_boxplot() +
  geom_jitter(
    width = 0.1,
    height = 0,
    alpha = 0.4
  ) +
  theme_classic()

p_roi_data
```

```{r}
p_roi_data2 <- roi_data %>% 
  filter(organism == "T. hydrogeniphilus") %>% 
  filter(label_at2H == 50) %>% 
  ggplot(
    aes(
      x = filename,
      y = Ratio_2Hx1H,
      color = filename
    )
  ) +
  geom_boxplot() +
  geom_point() +
  theme_classic()

p_roi_data2
```

```{r}
p_roi_data3 <- roi_data2 %>% 
  filter(organism == "T. hydrogeniphilus") %>% 
  filter(filename != "Thy_50_12_r4_f0_ratio2H-x-1H.tsv",
         filename != "Thy_50_12_r7_r_f0_ratio2H-x_1H.tsv") %>% 
  mutate(
    label_at2H = case_when(
      label_at2H == 0 ~ 0.00015576323,
      TRUE ~ label_at2H)
  ) %>% 
  ggplot(
    aes(
      x = label_at2H,
      y = Ratio_2Hx1H,
      color = organism
    )
  ) +
  # geom_boxplot(
  #   aes(
  #     group = label_at2H
  #   ),
  #   width = 0.5,
  #   outlier.shape = NA,
  #   color = "black"
  # ) +
  geom_jitter(
    width = 0.25,
    height = 0,
    alpha = 0.4,
    size = 0.5
  ) +
  ggdist::stat_halfeye(
    position = position_nudge(x = 2),
    color = "black"
  ) +
  theme_classic()

p_roi_data3
```

### ROI Counts (Raw)

```{r ion counts histogram} 
roi_data_joined %>% 
  ggplot(
    aes(x = `1H`)
  ) +
  geom_histogram() +
  labs(
    x = latex2exp::TeX("$^1H$ Counts (Raw)"),
    y = "ROI Count"
  ) +
  theme_minimal()

roi_data_joined %>% 
  ggplot(
    aes(x = `2H`)
  ) +
  geom_histogram() +
  labs(
    x = latex2exp::TeX("$^2H$ Counts (Raw)"),
    y = "ROI Count"
  ) +
  theme_minimal()
```

### 2H_counts versus 2R

```{r}
roi_data_joined %>% 
  pivot_longer(
    cols = c(`1H`, `2H`, SE), 
    values_to = "counts", 
    names_to = "channel"
    ) %>% 
  filter(channel != "SE") %>% 
  ggplot(
    aes(
      y = counts,
      x = `Ratio_2Hx1H`,
      color = channel
      )
  ) +
  geom_point(
    alpha = 0.5
  ) +
  facet_wrap(vars(channel), scales = "free") +
  scale_color_viridis_d(begin = 0.4, end = 0.8) +
  theme_classic() +
  theme(
    strip.background = element_blank(),
    strip.text = element_blank()
  )
```

### Area versus 2R

```{r}
roi_data_joined %>% 
  ggplot(
    aes(x = area,
        y = `Ratio_2Hx1H`,
        color = nanoSIMS_file_id,
        shape = organism)
  ) +
  geom_point() +
  coord_cartesian(xlim = c(0,2500)) +
  scale_color_viridis_d() +
  theme(
    legend.position = "NA"
  )
#plotly::ggplotly()
```

### Area versus 1H and 2H

```{r}
roi_data_joined %>% 
  pivot_longer(cols = c(`1H`, `2H`, SE), values_to = "counts", names_to = "channel") %>% 
  filter(channel != "SE") %>% 
  ggplot(aes(
    x = area,
    y = counts,
    color = channel
  )) +
  geom_point(
    alpha = 0.5,
    size = 1
  ) +
  geom_smooth(
    method = "lm",
    formula = "y~x",
    linetype = "dashed",
    size = 0.5,
    se = FALSE
  ) +
  scale_color_viridis_d(begin = 0.4, end = 0.8) +
  theme_classic()
```

### Deadtime correction effect

```{r}
p1H_dtc <- roi_data_joined %>% 
  ggplot(
    aes(
      x = `1H`,
      y = `1H_dtc`
      )
  ) +
  geom_abline(
    slope = 1,
    color = "red"
  ) +
  geom_point(
    alpha = 0.5
  ) +
  labs(
    title = latex2exp::TeX("$^1H$"),
    x = latex2exp::TeX("$C_{raw}$"),
    y = latex2exp::TeX("$C_{dtc}$")
  ) +
  theme_bw() +
  theme(
    plot.title = element_text(hjust = 0.5)
  )

p2H_dtc <- roi_data_joined %>% 
  ggplot(
    aes(
      x = `2H`,
      y = `2H_dtc`
      )
  ) +
  geom_abline(
    slope = 1,
    color = "red"
  ) +
  geom_point(
    alpha = 0.5
  ) +
  labs(
    title = latex2exp::TeX("$^2H$"),
    x = latex2exp::TeX("$C_{raw}$"),
    y = latex2exp::TeX("$C_{dtc}$")
  ) +
  theme_bw() +
  theme(
    plot.title = element_text(hjust = 0.5)
  )

p2R_dtc <- roi_data_joined %>% 
  ggplot(
    aes(
      x = `Ratio_2Hx1H2`,
      y = `Ratio_2Hx1H_dtc`
      )
  ) +
  geom_abline(
    slope = 1,
    color = "red"
  ) +
  geom_point(
    alpha = 0.5
  ) +
  labs(
    title = latex2exp::TeX("$^2R$"),
    x = latex2exp::TeX("$^2R_{raw}$"),
    y = latex2exp::TeX("$^2R_{dtc}$")
  ) +
  theme_bw() +
  theme(
    plot.title = element_text(hjust = 0.5)
  )

p2F_dtc <- roi_data_joined %>% 
  ggplot(
    aes(
      x = frac_2H,
      y = frac_2H_dtc
      )
  ) +
  geom_abline(
    slope = 1,
    color = "red"
  ) +
  geom_point(
    alpha = 0.5
  ) +
  labs(
    title = latex2exp::TeX("$^2F$"),
    x = latex2exp::TeX("$^2F_{raw}$"),
    y = latex2exp::TeX("$^2F_{dtc}$")
  ) +
  theme_bw() +
  theme(
    plot.title = element_text(hjust = 0.5)
  )

cowplot::plot_grid(p1H_dtc, p2H_dtc, p2R_dtc, p2F_dtc, nrow = 2)
```

```{r}
p2R_2 <- roi_data_joined %>% 
  ggplot(
    aes(
      x = Ratio_2Hx1H,
      y = Ratio_2Hx1H_2
      )
  ) +
  geom_abline(
    slope = 1,
    color = "red"
  ) +
  geom_point(
    alpha = 0.5
  ) +
  labs(
    title = latex2exp::TeX("$^2R$"),
    x = latex2exp::TeX("$^2R_{raw}$"),
    y = latex2exp::TeX("$^2R_{calculated}$")
  ) +
  theme_bw() +
  theme(
    plot.title = element_text(hjust = 0.5)
  )
p2R_2
```
