==> 01_Peak_Fitting_Data_Reduction.Rmd <==
---
title: "01 Peak Fitting Data Reduction"
author: "Tristan Caro"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Setup

## Libraries
```{r}
library(tidyverse)
library(readxl)
library(openxlsx)
library(magrittr)
```

## Function definitions

```{r}
select_fit_params <- function(df) {
  # the deparse(substitute(x)) changes the argument into a string that can be used
  sample_id_str = deparse(substitute(df))
  df %>% 
    select(fit, p, a , g, w) %>% 
    # add cell ids:
    mutate(sample = sample_id_str)
}

select_fit_areas <- function(df) {
  # the deparse(substitute(x)) changes the argument into a string that can be used
  sample_id_str = deparse(substitute(df))
  df %>% 
    # there are two "cell" columns - pick the first one
    select(unique(colnames(.))) %>% 
    select(cell, CD, contains("CH"), `CD%`) %>% 
    mutate(sample = sample_id_str) %>% 
    # remove NA rows introduced from fit dataset
    filter(!is.na(cell))
}

select_fit_xy <- function(df) {
  # change the argument into a string that will be used later
  sample_id_str = deparse(substitute(df))
  df %>% 
    select(unique(colnames(.))) %>% 
    # exclude columns from the other datasets
    select(-c(contains("CH"), contains("Peak2"), CD, `CD%`, p, a , g, w, fit)) %>% 
    # tidy it! pivot the dataset in order to change column names to x values
    pivot_longer(cols = -c(cell),
      values_to = "intensity",
      names_to = "wavenumber_cm") %>%
    # change these values to numeric, if not already
    mutate(
      wavenumber_cm = as.numeric(wavenumber_cm),
      intensity = as.numeric(intensity)
    ) %>% 
    mutate(sample = sample_id_str)
}

clean_area_data <- function(df) {
    df %>% mutate(
    cell_id = paste0(
      sample,
      "_", 
      sprintf("%02d", cell) # keep the cell id digits in 2 digits!
      ),
    f_l = case_when(
      sample == "Mb_40_y" ~ 40,
      sample == "Mb_20_x" ~ 20,
      TRUE ~ as.numeric(
        str_sub(
          cell_id, 
          start = -8, 
          end = -7)
        )
      )
    ) %>% 
  # remove the cell column
  select(-cell) %>% 
  # reorder the columns
  relocate(sample, cell_id)
}

clean_fit_data <- function(df) {
  df %>% 
    mutate(
    cell_id = paste0(
      sample,
      "_", 
      sprintf("%02d", cell) # keep the cell id digits in 2 digits!
      )
    ) %>% 
  # remove the cell column
  select(-cell) %>% 
  # reorder the columns
  relocate(sample, cell_id)
}
```

# Read data

## Read CD-2F dataset

```{r}
# Define where our large xlsx file is
path <- "data/Raman/fitted_data/CD_2F_RamanFits_FINAL.xlsx"

# Use openxlsx to get the names of the sheets
sheets <- openxlsx::getSheetNames(path)

# Read in all the data by lapply-ing the read.xlsx function across all the sheet names
fitted_data <- lapply(sheets, openxlsx::read.xlsx, xlsxFile=path)

# Assign names to the data frame
names(fitted_data) <- sheets
```

## Read time series test dataset

```{r}
# Define where our large xlsx file is
path2 <- "data/Raman/fitted_data/CD_RamanFits_GrowthExpt_TESTDATA.xlsx"

# use openxlsx to get the names of the sheets
sheets2 <- openxlsx::getSheetNames(path2)

# Read in all the data by lapply-ing the read.xlsx function across all the sheet names
cd20_test_dataset <- lapply(sheets2, openxlsx::read.xlsx, xlsxFile=path2)

# Assign names to the dataframe
names(cd20_test_dataset) <- sheets2

```


# Reduce Data

## Pull excel sheets into their own dataframes

I could probably do this more elegantly, but for now this will suffice.

For the CD2F dataset:

```{r}
Thy_00_07 <- fitted_data$`Thy_00_07!`
Thy_10_08 <- fitted_data$`Thy_10_08!`
Thy_20_09 <- fitted_data$`Thy_20_09!`
Thy_30_10 <- fitted_data$`Thy_30_10!`
Thy_40_11 <- fitted_data$`Thy_40_11!`
Thy_50_12 <- fitted_data$`Thy_50_12!`

Mb_00_13 <- fitted_data$`Mb_00_13!`
Mb_10_14 <- fitted_data$`Mb_10_14!`
Mb_20_23 <- fitted_data$`Mb_20_23!`
Mb_30_27 <- fitted_data$`Mb_30_27!`
Mb_40_y <- fitted_data$`Mb_40_y!`
Mb_50_18 <- fitted_data$`Mb_50_18!`

```

For the time-series CD20 dataset:

```{r}
Thy_20_T1_17h <- cd20_test_dataset$Thy_20_T1_17h
Thy_20_T2_24h <- cd20_test_dataset$Thy_20_T2_24h
Thy_20_T3_41h <- cd20_test_dataset$Thy_20_T3_41h
Thy_20_T4_48h <- cd20_test_dataset$Thy_20_T4_48h
```


## Separate the data

Here we separate the fit parameters, peak area summaries, and fitted peak paths (xy data) into their own dataframes, identifying them by the sample name.

For the CD2F dataset:
```{r}
fit_data <- bind_rows(
  select_fit_params(Thy_00_07),
  select_fit_params(Thy_10_08),
  select_fit_params(Thy_20_09),
  select_fit_params(Thy_30_10),
  select_fit_params(Thy_40_11),
  select_fit_params(Thy_50_12),
  select_fit_params(Mb_00_13),
  select_fit_params(Mb_10_14),
  select_fit_params(Mb_20_23),
  select_fit_params(Mb_30_27),
  select_fit_params(Mb_40_y),
  select_fit_params(Mb_50_18)
)

area_data <- bind_rows(
  select_fit_areas(Thy_00_07),
  select_fit_areas(Thy_10_08),
  select_fit_areas(Thy_20_09),
  select_fit_areas(Thy_30_10),
  select_fit_areas(Thy_40_11),
  select_fit_areas(Thy_50_12),
  select_fit_areas(Mb_00_13),
  select_fit_areas(Mb_10_14),
  select_fit_areas(Mb_20_23),
  select_fit_areas(Mb_30_27),
  select_fit_areas(Mb_40_y),
  select_fit_areas(Mb_50_18)
) %>% 
  clean_area_data()



peak_data_xy <- bind_rows(
  select_fit_xy(Thy_00_07),
  select_fit_xy(Thy_10_08),
  select_fit_xy(Thy_20_09),
  select_fit_xy(Thy_30_10),
  select_fit_xy(Thy_40_11),
  select_fit_xy(Thy_50_12),
  select_fit_xy(Mb_00_13),
  select_fit_xy(Mb_10_14),
  select_fit_xy(Mb_20_23),
  select_fit_xy(Mb_30_27),
  select_fit_xy(Mb_40_y),
  select_fit_xy(Mb_50_18)
) %>% 
  clean_fit_data()
```


For the CD20 test dataset:

```{r}
fit_data_cd20 <- bind_rows(
  select_fit_params(Thy_20_T1_17h),
  select_fit_params(Thy_20_T2_24h),
  select_fit_params(Thy_20_T3_41h),
  select_fit_params(Thy_20_T4_48h)
)

area_data_cd20 <- bind_rows(
  select_fit_areas(Thy_20_T1_17h),
  select_fit_areas(Thy_20_T2_24h),
  select_fit_areas(Thy_20_T3_41h),
  select_fit_areas(Thy_20_T4_48h)
) %>% 
  clean_area_data() %>%
  # need to add in f_l manually and clean CH
  mutate(
    f_l = 20,
    CH = as.numeric(CH)
    )

peak_data_xy_cd20 <- bind_rows(
  select_fit_xy(Thy_20_T1_17h),
  select_fit_xy(Thy_20_T2_24h),
  select_fit_xy(Thy_20_T3_41h),
  select_fit_xy(Thy_20_T4_48h)
) %>% 
  clean_fit_data()


# Create abbreviated CD20 dataset

area_data_cd20_tidy <- area_data_cd20 %>% 
  select(
    sample,
    cell_id,
    f_l,
    `CD%`
  )
```


# Clear memory

```{r}
# Clear individual dfs
rm(
  Thy_00_07,
  Thy_10_08,
  Thy_20_09,
  Thy_30_10,
  Thy_40_11,
  Thy_50_12,
  Mb_00_13,
  Mb_10_14,
  Mb_20_23,
  Mb_30_27,
  Mb_40_y,
  Mb_50_18
  )

rm(
  Thy_20_T1_17h,
  Thy_20_T2_24h,
  Thy_20_T3_41h,
  Thy_20_T4_48h
  )

# Clear the large lists
rm(
  fitted_data,
  cd20_test_dataset
  )
```


# Export

```{r}
# Export CD2F datasets
saveRDS(peak_data_xy, file = "cache/peak_data_xy.RDS")
saveRDS(area_data, file = "cache/area_data.RDS")
saveRDS(fit_data, file = "cache/fit_data.RDS")

# Export CD20 test datasets
saveRDS(peak_data_xy_cd20, file = "cache/peak_data_xy_cd20.RDS")
saveRDS(area_data_cd20, file = "cache/area_data_cd20.RDS")
saveRDS(fit_data_cd20, file = "cache/fit_data_cd20.RDS")
saveRDS(area_data_cd20_tidy, file = "cache/test_dataset.RDS")
```
