==> 02_CD2F_Co-Registration.Rmd <==
---
title: "02_CD2F_Co-Registration"
author: "Tristan Caro"
date: "`r Sys.Date()`"
output: html_document
---

# Setup


## Clear the environment

```{r}
rm(list=ls())
```


## Load libraries

```{r}
library(tidyverse)
```

## Read from cache
```{r}
# Import the nanoSIMS Data
roi_data_joined <- readRDS("cache/roi_data_joined.RDS")

# Import the CD% data (NOT poly-smoothed)
cd_pc_summarized <- readRDS("cache/cd_pc_data_summarized.RDS")

# Import the excluded cells list
excluded_cells <- readxl::read_excel("data/other/excluded_cells.xlsx") %>% pull(cell_id)

# Import the CD% data that results from our fitting in LabSpec6
cd_pc_summarized_fitted <- readRDS("cache/area_data.RDS")
```

## Join the data!

```{r}
co_registered <- roi_data_joined %>% 
  left_join(cd_pc_summarized,
            by = "cell_id") %>% 
  select(-c(`organism.y`)) %>% 
  # Exclude cells on the excluded cells list
  filter(!cell_id %in% excluded_cells) %>% 
  # Exclude samples on the excluded samples list
  filter(!sample %in% c("Mb_20_23"))

co_registered_fitted <- roi_data_joined %>% 
  left_join(cd_pc_summarized_fitted,
            by = "cell_id") %>% 
  # Exclude cells on the excluded cells list
  filter(!cell_id %in% excluded_cells) %>% 
  # Exclude samples on the excluded samples list
  filter(!sample %in% c("Mb_20_23"))
```

```{r}
saveRDS(co_registered, "cache/co_registered.RDS")
saveRDS(co_registered_fitted, "cache/co_registered_fitted.RDS")
```
