==> 03_CD2F_Sensitivity.Rmd <==
---
title: "R Notebook"
author: "Tristan Caro"
date: "`r Sys.Date()`"
output: html_notebook
editor_options: 
  chunk_output_type: console
---

# Setup

## Clear ENV
```{r}
rm(list=ls())
```

## Libraries

```{r}
library(tidyverse)
library(geomtextpath)
library(plotly)
library(patchwork)
library(ggpp)
```

## Functions

```{r}
calculate_mu <- function(a, F2, F1, FL, t2, t1 = 0) {
  return(- (1/(t2 - t1)) * (log((a * FL - F2)/(a*FL - F1))))
}

calculate_sigma_mu <- function(a, F2, F1, FL, t2, t1 = 0, sigma_a = 0, sigma_F2 = 0, sigma_F1 = 0, sigma_FL = 0) {
  return(
    sqrt(
      ((a*FL - F2) * sigma_F1)^2 +
        
      ((a*FL - F1) * sigma_F2)^2 +
        
      (a * (F1 - F2) * sigma_FL)^2 +
        
      (FL * (F1 - F2) * sigma_a)^2
    ) / ( 
      (t2-t1) * (a * FL - F1) * (a * FL - F2) 
    )
  )
}

```



# Sensitivity Calculations

```{r}
test_dataset <- tibble(
  a = 1, # assimilation factor is set by the organism: 0.878 for Mb, 0.761 for Thy
  F1 = 0, # ppm
  # errors
  sigma_a = 0.020, # uncertainty in A is SE of Mb `a`: 0.020
  sigma_FL = 0, # uncertainty in FL is assumed to be zero for high isotopic abundance
  sigma_F1 = 0, # at%: SE of the method blank
  sigma_F2 = 2.888494 # at%: average SE of proxy
  ) %>% 
  crossing(
  FL = c(5, 10, 15, 20, 25, 30, 35, 40, 45, 50),
  dt = c(1, 5, 7, 30, 60, 100),
  frac_F2 = seq(-5, 0, by = 0.01) %>% exp() # microbial fractional progression to FL
  ) %>% 
  mutate(
   F2 = F1 + FL * frac_F2, # model F2 values: at% 
   mu.d = calculate_mu(a = a, F2 = F2, F1 = F1, FL = FL, t2 = dt),
   TD.days = log(2) / mu.d,
   sigma_mu.d = calculate_sigma_mu(
      a = a, F2 = F2, F1 = F1, FL = FL, t2 = dt,
      sigma_a = sigma_a, 
      sigma_F2 = sigma_F2, 
      sigma_F1 = sigma_F1, 
      sigma_FL = sigma_FL
    ),
    rel_error = sigma_mu.d / mu.d
  ) %>%
  filter(!is.na(mu.d), !is.na(rel_error))


```



# Define Plotting Functions

## RE vs. gen_time

```{r}
plot_rel_error <- function(test_dataset, d_t, f_label, xlimits, include_legend = TRUE, include_caption = TRUE) {
  upper_lim <- test_dataset |> 
    filter(dt == d_t) |> 
    filter(rel_error <= 0.5) |>
    filter(FL == f_label) |> 
    pull(TD.days) |> 
    min()
  
  upper_lim_mu_d <- test_dataset |> 
    filter(dt == d_t) |> 
    filter(rel_error <= 0.5) |>
    filter(FL == f_label) |> 
    pull(mu.d) |> 
    min()
  
  lower_lim <- test_dataset |> 
    filter(dt == d_t) |> 
    filter(rel_error <= 0.5) |>
    filter(FL == f_label) |> 
    pull(TD.days) |> 
    max()
  
  lower_lim_mu_d <- test_dataset |> 
    filter(dt == d_t) |> 
    filter(rel_error <= 0.5) |>
    filter(FL == f_label) |> 
    pull(mu.d) |> 
    max()
  
  # find the error minima (optima)
  minimums <- test_dataset |> 
    filter(dt == d_t) |> 
    group_by(FL) |> 
    filter(rel_error == min(rel_error))
  
  minimum_label <- minimums |>
    filter(FL > 15) |> 
    filter(FL == f_label) |> 
    pull(TD.days)
  
  test_dataset %>%
    # Argument input here to filter dataset to correct incubation time:
    filter(dt == d_t) |> 
    filter(FL > 15) |> 
    ggplot() +
    aes(
      x = TD.days,
      y = rel_error,
      color = as.factor(FL)
    ) +
    annotate(
      geom = "segment", x = lower_lim, xend = lower_lim, 
      y = 0, yend = 0.5, color = "gray"
      ) +
    annotate(
      geom = "segment", x = upper_lim, xend = upper_lim, 
      y = 0, yend = 0.5, color = "gray"
      ) +
    annotation_custom(grid::textGrob(paste("dt =", d_t, "days"), 0.80, 0.2)) +
    geom_segment(
      data = minimums |> filter(FL > 15) |> filter(FL == f_label),
      aes(
        x = TD.days,
        xend = TD.days,
        y = 0,
        yend = 50,
        color = as.factor(FL)
      ),
      alpha = 0.7
    ) +
    geom_line(linewidth = 1) +
    geom_point(
      data = minimums |> filter(FL > 15),
      aes(fill = as.factor(FL)),
      shape = 21,
      color = "black",
      show.legend = FALSE,
      size = 2.5,
      stroke = 1
    ) +
    scale_color_viridis_d(option = "plasma", begin = 0.1, end = 0.9) +
    scale_fill_viridis_d(option = "plasma", begin = 0.1, end = 0.9) +
    # Secondary axis with ticks marked
    scale_x_continuous(
      sec.axis = dup_axis(
        breaks = c(round(upper_lim, 1), round(minimum_label, 1), round(lower_lim, 1))
      )
    ) +
    scale_y_continuous(labels = scales::percent) + 
    coord_cartesian(
      ylim = c(0, 0.5),
      xlim = xlimits,
      expand = FALSE
    ) +
    ggprism::annotation_ticks(sides = "bl") +
    labs(
      x = "Generation Time (Days)",
      y = "Relative Error",
      color = latex2exp::TeX("Label Enrichment ($^{2}F_L$ at. %)"),
      caption = if_else(include_caption, paste(
        "Quantification of growth: \n",
        "Generation times: between", round(upper_lim, 1), "and",
        round(lower_lim, 1), "days, \n",
        "Growth rate: between", round(lower_lim_mu_d, 3), "and",
        round(upper_lim_mu_d, 3), "days ^-1"),"")
    ) +
    theme_bw() +
    theme(
      legend.position = if_else(include_legend, "bottom", "NA"),
      axis.text.x.top = element_text(angle = 45, hjust = 0),
      axis.title.x.top = element_blank(),
      panel.grid.major = element_blank(),
      panel.grid.minor = element_blank(),
      axis.text = element_text(color = "black", face = "bold")
      #panel.border = element_rect(color = "black", size = 2, fill = NA),
      #axis.ticks = element_line(linewidth = 1)
    )
}
 
# TEST THE FUNCTION
plot_rel_error(
  test_dataset = test_dataset,
  d_t = 30,
  xlimits = c(0,175),
  f_label = 35,
  include_legend = TRUE,
  include_caption = TRUE
)
```

## RE Vs. µ [deprecated]

```{r}
plot_rel_error_mu <- function(test_dataset, d_t, f_label, xlimits, include_legend = TRUE, include_caption = TRUE) {
  upper_lim <- test_dataset |> 
    filter(dt == d_t) |> 
    filter(rel_error <= 0.5) |>
    filter(FL == f_label) |> 
    pull(TD.days) |> 
    min()
  
  upper_lim_mu_d <- test_dataset |> 
    filter(dt == d_t) |> 
    filter(rel_error <= 0.5) |>
    filter(FL == f_label) |> 
    pull(mu.d) |> 
    min()
  
  lower_lim <- test_dataset |> 
    filter(dt == d_t) |> 
    filter(rel_error <= 0.5) |>
    filter(FL == f_label) |> 
    pull(TD.days) |> 
    max()
  
  lower_lim_mu_d <- test_dataset |> 
    filter(dt == d_t) |> 
    filter(rel_error <= 0.5) |>
    filter(FL == f_label) |> 
    pull(mu.d) |> 
    max()
  
  test_dataset %>%
    # Argument input here to filter dataset to correct incubation time:
    filter(dt == d_t) |> 
    filter(FL > 15) |> 
    ggplot() +
    aes(
      x = TD.days,
      y = rel_error,
      color = as.factor(FL)
    ) +
    annotate(geom = "segment", x = lower_lim, xend = lower_lim, y = 0, yend = 0.5, color = "gray") +
    annotate(geom = "segment", x = upper_lim, xend = upper_lim, y = 0, yend = 0.5, color = "gray") +
    #annotate(geom = "label", x = upper_lim, y = 0.1, color = "black", label = paste0(round(upper_lim, 1), " d")) +
    #annotate(geom = "label", x = lower_lim, y = 0.1, color = "black", label = paste0(round(lower_lim, 1), " d")) +
    geom_line(linewidth = 1) +
    scale_color_viridis_d(option = "plasma", begin = 0.1, end = 0.9) +
    # Secondary axis with ticks marked
    scale_x_continuous(
      sec.axis = dup_axis(
        breaks = c(round(upper_lim, 1), round(lower_lim, 1))
      )
    ) +
    # Secondary axis with growth rate:
    # scale_x_continuous(
    #   sec.axis = sec_axis(~ log(2)/., breaks = c(0.1, 0.01, 0.002)),
    #   
    #   ) +
    scale_y_continuous(labels = scales::percent) + 
    coord_cartesian(
      ylim = c(0, 0.5),
      xlim = xlimits,
      expand = FALSE
    ) +
    labs(
      x = "Generation Time (Days)",
      y = latex2exp::TeX("Relative Error (%) =     $\\frac{\\sigma_{\\mu}}{\\mu}$"),
      color = latex2exp::TeX("Tracer Strength $^{2}H \\%$"),
      title = paste(d_t, "day incubation time"),
      caption = if_else(include_caption, paste(
        "Quantification of growth: \n",
        "Generation times: between", round(upper_lim, 1), "and", round(lower_lim, 1), "days, \n",
        "Growth rate: between", round(lower_lim_mu_d, 3), "and", round(upper_lim_mu_d, 3), "days ^-1"
        ),
        "")
    ) +
    theme_classic() +
    theme(
      legend.position = if_else(include_legend, "bottom", "NA"),
      axis.text.x.top = element_text(angle = 45, hjust = 0),
      axis.title.x.top = element_blank(),
      #panel.border = element_rect(color = "black", size = 2, fill = NA),
      #axis.ticks = element_line(linewidth = 1)
    )
}
```

## RE vs. 2F

```{r}
plot_rel_error_at2H <- function(test_dataset, d_t, f_label, xlimits, include_legend = TRUE, include_caption = TRUE) {
  
  # find upper lim 2F
  upper_lim <- test_dataset |> 
    filter(dt == d_t) |> 
    filter(rel_error <= 0.5) |>
    filter(FL == f_label) |> 
    pull(F2) |> 
    min()
  
  # find upper lim growth rate
  upper_lim_mu_d <- test_dataset |> 
    filter(dt == d_t) |> 
    filter(rel_error <= 0.5) |>
    filter(FL == f_label) |> 
    pull(mu.d) |> 
    min()
  
  # find lower lim 2F
  lower_lim <- test_dataset |> 
    filter(dt == d_t) |> 
    filter(rel_error <= 0.5) |>
    filter(FL == f_label) |> 
    pull(F2) |> 
    max()
  
  # find lower lim growth rate
  lower_lim_mu_d <- test_dataset |> 
    filter(dt == d_t) |> 
    filter(rel_error <= 0.5) |>
    filter(FL == f_label) |> 
    pull(mu.d) |> 
    max()
  
  # find the error minima (optima)
  minimums <- test_dataset |> 
    filter(dt == d_t) |> 
    group_by(FL) |> 
    filter(rel_error == min(rel_error))
  
  minimum_label <- minimums |>
    filter(FL > 15) |> 
    filter(FL == f_label) |> 
    pull(F2)

  
  test_dataset %>%
    # Argument input here to filter dataset to correct incubation time:
    filter(dt == d_t) |> 
    filter(FL > 15) |> 
    ggplot() +
    aes(
      x = F2,
      y = rel_error,
      color = as.factor(FL)
    ) +
    annotate(
      geom = "segment", x = lower_lim, xend = lower_lim,
      y = 0, yend = 0.5, color = "gray") +
    annotate(
      geom = "segment", 
      x = upper_lim, xend = upper_lim, 
      y = 0, yend = 0.5, color = "gray") +
    geom_line(linewidth = 1) +
    geom_segment(
      data = minimums |> filter(FL > 15) |> filter(FL == f_label),
      aes(
        x = F2,
        xend = F2,
        y = 0,
        yend = 50,
        color = as.factor(FL)
      ),
      alpha = 0.7
    ) +
    geom_point(
      data = minimums |> filter(FL > 15),
      aes(fill = as.factor(FL)),
      shape = 21,
      color = "black",
      show.legend = FALSE,
      size = 2.5,
      stroke = 1
    ) +
    scale_color_viridis_d(end = 0.9) +
    scale_fill_viridis_d(end = 0.9) +
    # Secondary axis with ticks marked
    scale_x_continuous(
      sec.axis = dup_axis(
        breaks = c(round(upper_lim, 1), 
                   round(minimum_label, 1), 
                   round(lower_lim, 1)),
        labels = scales::label_percent(scale = 1)
      ),
      labels = scales::label_percent(scale = 1)
    ) +
    scale_y_continuous(labels = scales::label_percent()) + 
    coord_cartesian(
      ylim = c(0, 0.5),
      xlim = xlimits,
      expand = FALSE
    ) +
    ggprism::annotation_ticks(sides = "bl") +
    labs(
      x = latex2exp::TeX("$^2F$ (at. %)"),
      y = "Relative Error",
      color = latex2exp::TeX("Label Enrichment ($^{2}F_L$ at. %)"),
      title = latex2exp::TeX("$^2F$ quantification range"),
      caption = if_else(include_caption, paste(
        "Range of quantification: \n",
        "CD% values: ", round(upper_lim, 1), "-",
        round(lower_lim, 1), "%, \n",
        "Growth rate: ", round(lower_lim_mu_d, 3), "-", 
        round(upper_lim_mu_d, 3), "days^-1"),"")
    ) +
    theme_bw() +
    theme(
      panel.grid = element_blank(),
      axis.text = element_text(color = "black", face = "bold"),
      legend.position = if_else(include_legend, "bottom", "NA"),
      axis.text.x.top = element_text(angle = 45, hjust = 0),
      axis.title.x.top = element_blank(),
      plot.margin = margin(0.5,0.5,0.5,0.5, "cm")
    )
}
 

plot_rel_error_at2H(
  test_dataset = test_dataset,
  d_t = 30,
  xlimits = c(0,50),
  f_label = 35,
  include_legend = TRUE,
  include_caption = TRUE
)

```

# Generate Nice Plots
```{r}

# range of quant
panel_a <- plot_rel_error_at2H(
  test_dataset = test_dataset,
  d_t = 30,
  xlimits = c(0,50),
  f_label = 35,
  include_legend = TRUE,
  include_caption = FALSE
)

# dt = 5 days
panel_b <- plot_rel_error(
  test_dataset = test_dataset,
  d_t = 5,
  xlimits = c(0,50),
  f_label = 35,
  include_legend = FALSE,
  include_caption = FALSE
)

# dt = 30 days
panel_c <- plot_rel_error(
  test_dataset = test_dataset,
  d_t = 30,
  xlimits = c(0,175),
  f_label = 35,
  include_legend = TRUE,
  include_caption = FALSE
)


combined <- panel_a + (panel_b / panel_c)

combined

cowplot::save_plot(
  plot = combined,
  filename = "fig_output/figure3.pdf",
  base_height = 7,
  base_width = 12
)
 
```

# Supp Fig. Multiple dt

```{r}
dt1 <- plot_rel_error(
  test_dataset = test_dataset,
  d_t = 1,
  xlimits = c(0,7),
  f_label = 35,
  include_legend = FALSE,
  include_caption = FALSE
)


dt5 <- plot_rel_error(
  test_dataset = test_dataset,
  d_t = 5,
  xlimits = c(0,45),
  f_label = 35,
  include_legend = FALSE,
  include_caption = FALSE
)



dt7 <- plot_rel_error(
  test_dataset = test_dataset,
  d_t = 7,
  xlimits = c(0,45),
  f_label = 35,
  include_legend = FALSE,
  include_caption = FALSE
)


dt30 <- plot_rel_error(
  test_dataset = test_dataset,
  d_t = 30,
  xlimits = c(0,205),
  f_label = 35,
  include_legend = FALSE,
  include_caption = FALSE
)


dt60 <- plot_rel_error(
  test_dataset = test_dataset,
  d_t = 60,
  xlimits = c(0,350),
  f_label = 35,
  include_legend = FALSE,
  include_caption = FALSE
)


dt100 <- plot_rel_error(
  test_dataset = test_dataset,
  d_t = 100,
  xlimits = c(0,550),
  f_label = 35,
  include_legend = FALSE,
  include_caption = FALSE
)

p_supp_dt <- cowplot::plot_grid(
  dt1, dt5, dt7, dt30, dt60, dt100,
  ncol = 2
)
p_supp_dt

cowplot::save_plot(
  filename = "fig_output/supp_dt.pdf",
  plot = p_supp_dt,
  base_height = 10,
  base_width = 12
)

```

# FT vs. t

```{r}
calculate_FT <- function(a = 1, FL, mu, t, F0) {
  e = exp(1) # define euler's number
  return(
    a * FL * (1 - e^(-mu * t)) + F0 * e^(-mu * t)
  )
}

```

```{r}
test_dataset_2 <- 
  crossing(
    a = c(0.1, 0.3, 0.5, 0.7, 0.9),
    FL = c(10, 20, 30, 35, 40, 50),
    mu = c(0.05, 0.13),
    t = seq(0, 1000, by = 1)
  ) |> 
  mutate(
    FT = calculate_FT(
      a = a,
      FL = FL,
      F0 = 0,
      mu = mu,
      t = t
    ),
    aFL = a * FL
  )
```

```{r}
p_FT_vs_a <- test_dataset_2 |> 
  filter(FL == 35) |> 
  ggplot(
    aes(
      x = t,
      y = FT,
      color = as.factor(a)
    )
  ) +
  geom_hline(aes(yintercept = aFL, color = as.factor(a)), alpha = 0.3, linetype = "dashed") +
  geom_line() +
  facet_wrap(vars(as.factor(mu)), scales = "free_y") +
  coord_cartesian(xlim = c(0, 80), ylim = c(0, 33), expand = FALSE) +
  scale_color_viridis_d(option = "plasma", begin = 0, end = 0.8) +
  labs(
    x = latex2exp::TeX("\\Delta t"),
    y = latex2exp::TeX("$^{2}F_{biomass}$"),
    color = latex2exp::TeX("$a_{w}$"),
    linetype = latex2exp::TeX("$\\mu$")
  ) +
  theme_bw() +
  theme(
    panel.grid = element_blank(),
    strip.text = element_text(color = "white"),
    strip.background = element_rect(fill = "black"),
    axis.text = element_text(face = "bold", color = "black"),
    panel.spacing = unit(2, "lines"),
    legend.position = "bottom",
    aspect.ratio = 1
  )
p_FT_vs_a

cowplot::save_plot(
  plot = p_FT,
  filename = "fig_output/p_FT.pdf",
  base_height = 5,
  base_width = 7
)
```

```{r}
p_FT_vs_FL <- test_dataset_2 |> 
  filter(a == 0.7) |> 
  ggplot(
    aes(
      x = t,
      y = FT,
      color = as.factor(FL)
    )
  ) +
  geom_hline(aes(yintercept = aFL, color = as.factor(FL)), alpha = 0.3, linetype = "dashed") +
  geom_line() +
  facet_wrap(vars(as.factor(mu)), scales = "free_y") +
  coord_cartesian(xlim = c(0, 80), ylim = c(0, 40), expand = FALSE) +
  scale_color_manual(values = ggpubfigs::friendly_pal("bright_seven")) +
  labs(
    x = latex2exp::TeX("\\Delta t"),
    y = latex2exp::TeX("$^{2}F_{biomass}$"),
    color = latex2exp::TeX("$F_{L}$"),
    linetype = latex2exp::TeX("$\\mu$")
  ) +
  theme_bw() +
  theme(
    panel.grid = element_blank(),
    strip.text = element_text(color = "white"),
    strip.background = element_rect(fill = "black"),
    axis.text = element_text(face = "bold", color = "black"),
    panel.spacing = unit(2, "lines"),
    legend.position = "top",
    aspect.ratio = 1
  )
p_FT_vs_FL

cowplot::save_plot(
  plot = p_FT_vs_FL,
  filename = "fig_output/p_FT_vs_FL.pdf",
  base_height = 5,
  base_width = 5
)
```


```{r}
p_FT_schematic <- test_dataset_2 |> 
  filter(FL == 50, mu == 0.05, !a %in% c(0.1, 0.3, 0.5, 0.9)) |>
  ggplot(
    aes(
      x = t,
      y = FT,
      color = as.factor(a)
    )
  ) +
  geom_hline(
    aes(
      yintercept = aFL,
      color = as.factor(a)),
    alpha = 0.3, 
    linetype = "dashed"
    ) +
  geom_line(linewidth = 1.5) +
  coord_cartesian(xlim = c(0, 90), ylim = c(0, 33), expand = FALSE) +
  scale_color_viridis_d(option = "plasma", begin = 0.3, end = 0.8) +
  labs(
    x = latex2exp::TeX("\\Delta t"),
    y = latex2exp::TeX("$^{2}F_{biomass}$"),
    color = latex2exp::TeX("$a_{w}$"),
    linetype = latex2exp::TeX("$\\mu$")
  ) +
  theme_bw() +
  theme(
    axis.text = element_blank(),
    axis.ticks = element_blank(),
    panel.grid = element_blank(),
    aspect.ratio = 1,
    legend.position = "none"
  )
p_FT_schematic

cowplot::save_plot(
  plot = p_FT_schematic,
  filename = "fig_output/p_FT_schematic.pdf",
  base_height = 4,
  base_width = 4
)
```



# Assess multiple sources of uncertainty
```{r}

# error matrix
error_matrix <-
  tribble(
     ~use_a, ~use_FL, ~use_F1, ~use_F2, ~panel,
     1,       1,      1,       1,       "all errors",
     1,       0,      0,       0,       "a error only",
     0,       1,      0,       0,       "FL error only",
     0,       0,      1,       0,       "F1 error only",
     0,       0,      0,       1,       "F2 error only"
  ) %>%
  mutate(panel = as_factor(panel))

# constants
model_data <- 
  tibble(
  a = 1, # assimilation factor is 1: assume autotrophic growth
  t.days = 7, # days (doesn't actually matter, all cancels)
  F1 = 0, # ppm
  # errors
  sigma_a = 0.02, # no uncertainty in alpha because defined cultures
  sigma_FL = 0, # uncertainty in FL is unknown: determine with pipetting?
  sigma_F1 = 0, # at%
  sigma_F2 = 3.5 # at%
) %>%
  # dynamics
  crossing(
    FL = c(5, 10, 20, 30, 40, 50, 100), # label strength: at%
    F2_interval = seq(-100, 0, by = 0.01) %>% exp() # fraction of F2
  ) %>%
  # mu calculations
  mutate(
    F2 = F1 + FL * F2_interval, # model F2 values: at%
    t_factor = paste(t.days, "days") %>% as_factor(),
    mu.1_day = calculate_mu(a = a, F2 = F2, F1 = F1, FL = FL, t2 = t.days),
    TD.days = log(2) / mu.1_day
  ) %>%
  # error calculations
  crossing(error_matrix) %>%
  mutate(
    sigma_mu.1_day = calculate_sigma_mu(
      a = a, F2 = F2, F1 = F1, FL = FL, t2 = t.days,
      sigma_a = use_a * sigma_a, 
      sigma_F2 = use_F2 * sigma_F2, 
      sigma_F1 = use_F1 * sigma_F1, 
      sigma_FL = use_FL * sigma_FL
    ),
    rel_error = sigma_mu.1_day / mu.1_day
  ) %>%
  filter(!is.na(mu.1_day), !is.na(rel_error))

# Find minimum error condition:
error_minimum <- model_data %>% 
  filter(panel == "all errors") %>% 
  filter(FL == 30) %>% 
  filter(is.finite(rel_error)) %>% 
  filter(t.days == 7) %>% 
  filter(TD.days / 7 < 100,
         TD.days / 7 > 0.1) %>% 
  group_by(FL) %>%
  filter(rel_error == min(rel_error)) %>% 
  select(TD.days, t.days, rel_error) %>% 
  mutate(TD.days.t = TD.days / t.days)

error_min_td.t = error_minimum %>% pull(TD.days.t)
error_min_rel_error = error_minimum %>% pull(rel_error)
```

Plot the sources of uncertainty
```{r}
p_rel_error_all <- model_data %>%
  filter(panel == "all errors") %>% 
  ggplot() + 
  aes(
    x = TD.days / t.days, 
    y = rel_error, 
    color = as.factor(FL), 
    label = paste0(as.factor(FL), "%")
    ) + 
  geom_vline(
    xintercept = error_min_td.t, 
    color = "gray", 
    linetype = "dotted") +
  geom_line(size = 1.25) +
  geom_textline(
    hjust = "ymin",
    vjust = 1.5, 
    show.legend = FALSE
    ) +
  scale_x_log10(
    position = "top",
    breaks = c(0.1, 0.25, 0.5, 1, 2.5, 5, 10, 25)
    ) +
  scale_y_continuous(labels = scales::percent) + 
  scale_color_viridis_d(option = "plasma", begin = 0, end = 0.8) +
  coord_cartesian(ylim = c(0, 0.5), xlim = c(0.1, 25), expand = FALSE) +
  labs(
    x = latex2exp::TeX("$\\frac{T_G}{\\Delta t}$"), 
    y = latex2exp::TeX("$\\frac{\\sigma_{\\mu}}{\\mu}$"),
    color = "Label Strength (%)"
  ) + 
  #facet_wrap(~panel) +
  theme_bw() +
  theme(
    axis.text.x = element_text(angle = 45, hjust = 0),
    legend.position = "bottom"
  )

p_rel_error_all
# 
# cowplot::save_plot(
#   plot = p_rel_error_all,
#   filename = "fig_output/rel_error_all.pdf",
#   base_height = 5,
#   base_width = 5
# )
```


# Plotly

```{r}
plotly_rel_error_at2H <- function(test_dataset, d_t, xlimits = c(0, 50)) {
  
  plot <- test_dataset %>%
    # Argument input here to filter dataset to correct incubation time:
    filter(dt == d_t) |> 
    filter(FL > 15) |> 
    ggplot() +
    aes(
      x = F2,
      y = rel_error,
      color = as.factor(FL)
    ) +
    geom_line(linewidth = 1) +
    geom_hline(yintercept = 0.5, linetype = "dotted") +
    scale_color_viridis_d(end = 0.9) +
    # Set axis labels here. Labs will cause it to crash.
    scale_x_continuous(name = "Biomass (D at. %)") +
    scale_y_continuous(name = "Relative Error", labels = scales::percent) + 
    coord_cartesian(
      ylim = c(0, 0.5),
      xlim = xlimits,
      expand = FALSE
    ) +
    labs(
      title = paste(d_t, "day incubation time"),
    ) +
    theme_bw() +
    theme(
      legend.title = element_text("Tracer Strength (D at. %)")
    )
  ggplotly(plot) |> 
    layout(
    xaxis = list(
      title = TeX("^2F_{Biomass}")),
    yaxis = list(
      title = TeX("\\text{Relative Error } \\left(\\frac{\\sigma_{\\mu}}{\\mu} \\right)")),
    legend = list(
      title = list(text = TeX("^2F_{Tracer}")))
    ) |> 
    config(.Last.value, mathjax = 'cdn')
}


plotly_rel_error_at2H(test_dataset = test_dataset, d_t = 30)

```

```{r}
plotly_rel_error <- function(test_dataset, d_t, f_label, xlimits, include_legend = TRUE) {
  
  plot <- test_dataset %>%
    # Argument input here to filter dataset to correct incubation time:
    filter(dt == d_t) |> 
    filter(FL > 15) |> 
    ggplot() +
    aes(
      x = TD.days,
      y = rel_error,
      color = as.factor(FL)
    ) +
    geom_line(linewidth = 1) +
    geom_hline(yintercept = 0.5, linetype = "dotted") +
    scale_color_viridis_d(option = "plasma", begin = 0.1, end = 0.9) +
    # Secondary axis with ticks marked
    scale_x_continuous() +
    scale_y_continuous(labels = scales::percent) + 
    coord_cartesian(
      ylim = c(0, 0.5),
      xlim = xlimits,
      expand = FALSE
    ) +
    labs(
      color = "Tracer Strength (D at. %)",
      title = paste(d_t, "day incubation time"),
    ) +
    theme_classic() +
    theme()
  
  ggplotly(plot) |> 
    layout(
    xaxis = list(
      title = TeX("^2F_{Biomass}")),
    yaxis = list(
      title = TeX("\\text{Relative Error } \\left(\\frac{\\sigma_{\\mu}}{\\mu} \\right)")),
    legend = list(
      title = list(text = TeX("^2F_{Tracer}")))
    ) |> 
    config(.Last.value, mathjax = 'cdn')
}

plotly_rel_error(test_dataset = test_dataset, d_t = 30, f_label = 30, xlimits = c(0, 300))
```
