==> 04_molecules_H.qmd <==
---
= ---
title: "04_molecules_H"
format:
  html:
    code-fold: true
    code-summary: "Show the code"
editor: visual
editor_options: 
  chunk_output_type: console
bibliography: references.bib
production:
  CHROMOTE_CHROME: /Applications/Brave Browser.app
---

## Molecular Inventory of H

Here we generate a rough estimate of the fraction of cellular hydrogen that is exchangeable with the aqueous environment, versus non-exchangeable hydrogen.

### Setup

```{r}
#| include: false
library(tidyverse)
library(gt)
library(webshot2)
```

### Load data

```{r}
#| include: false
molecule_H <- readxl::read_excel("data/chemdraws/all_the_molecules.xlsx") |> 
  mutate(MW = as.numeric(MW)) |> 
  rename(umol_per_g = `amt (umol/g dried cells)`)
molecule_H


```

### Mole fraction of cell components

First let's examine the relative abundances of various biomolecules:

```{r}
table_gt <- molecule_H |> 
  select(compound, group, umol_per_g, H, H_exchangable, frac_H_exc) |> 
  rename(`umol per g` = umol_per_g,
         `Total H` = H,
         `H (Exchangeable)` = H_exchangable,
         `Fraction Exchangeable` = frac_H_exc) |> 
  gt() |> 
  data_color(columns = `umol per g`, palette = "Greens") |> 
  data_color(columns = `Fraction Exchangeable`, palette = "Blues") |> 
  tab_style(
    locations = cells_body(columns = compound),
    style = cell_text(weight = "bold")
  )
table_gt
```

Based on this dataset, the total $\mu mol \cdot g$ of biomolecules per cell mass is:

```{r}
total_umol_per_g <- molecule_H |> pull(umol_per_g) |> sum()
total_umol_per_g
```

so we can calculate the relative fraction of each biomolecule and plot it:

```{r}
molecule_H <- molecule_H |> 
  # this represents the fraction of this biomolecule relative to all the biomolecules
  mutate(molecule_mol_frac = umol_per_g / total_umol_per_g)
```

```{r}
#| fig-width: 6
#| fig-height: 8

p_mole_frac <- molecule_H |> 
  mutate(compound = fct_reorder(.f = compound, .x = molecule_mol_frac)) |> 
  ggplot() +
  aes(
    x = molecule_mol_frac,
    y = compound,
    fill = group
  ) +
  geom_col() +
  scale_x_continuous(labels = scales::percent) +
  labs(
    x = "Mole %"
  ) +
  theme_bw() +
  theme(
    axis.title.y = element_blank(),
    legend.title = element_blank(),
    panel.grid = element_blank(),
    axis.text = element_text(color = "black"),
    legend.position = "None"
  )

p_mole_frac
```

```{r}

```

### Calculate H

To calculate the mole fraction of hydrogen in a cell that is exchangeable, we first need to know (i) the total number of moles of H in a cell and (ii) the fraction of these that are exchangeable:

$$
\begin{aligned}
F_{H_{ex}} &= \frac{H_{ex}}{H_{ex} + H_{non}} \\
\space \\
H_{total} &= H_{ex} + H_{non} 
\end{aligned}
$$

where $H_{total}$ is the total moles of H in a cell, $H_{ex}$ is the moles of hydrogen in a cell that is bonded at an exchangeable site, $H_{non}$ is moles of hydrogen in a cell at a non-exchangeable site, and $F_{H_{ex}}$ is the fraction of cellular hydrogen that is exchangeable. From the Neinhardt dataset [@escheric1987], we have constraints on the mole fraction of many biocomponents:

```{r}
molecule_H_calc <- molecule_H |> 
  # calculate total number of moles H this component contributes to the cell
  # calculate total number of moles exchangeable H this component contributes to the cell
  mutate(
    umol_Htot_per_g = umol_per_g * H,
    umol_Hex_per_g = umol_per_g * H_exchangable,
    umol_Hnon_per_g = umol_Htot_per_g - umol_Hex_per_g
  )

molecule_H_calc |> 
  summarize(
    sum_umol_Htot_per_g = sum(umol_Htot_per_g),
    sum_umol_Hex_per_g = sum(umol_Hex_per_g)
  ) |> 
  mutate(frac_H_exchangeable = sum_umol_Hex_per_g / sum_umol_Htot_per_g) |> 
  rename(
    `µmol H (Exchangeable) / g` = sum_umol_Hex_per_g,
    `µmol H (Total) / g` = sum_umol_Htot_per_g,
    `Fraction of Exchangeable H` = frac_H_exchangeable
  ) |> 
  knitr::kable()
```

So, by tabulating all of the hydrogen in the cell and determining whether it is exchangeable or not, we can see that about **21%** of a cell's H pool is rapidly exchangeable with water.

```{r}
p_umol_g <- molecule_H_calc |> 
  select(compound, group, umol_Hex_per_g, umol_Hnon_per_g) |> 
  pivot_longer(cols = c(umol_Hnon_per_g, umol_Hex_per_g),
               names_to = "source", values_to = "umol_per_g") |> 
  mutate(
    source = if_else(
      source == "umol_Hnon_per_g",
      true = "Non-Exchangeable",
      false = "Exchangeable"
    )
  ) |> 
  mutate(compound = fct_reorder(.f = compound, .x = umol_per_g)) |> 
  ggplot() +
  aes(
    x = umol_per_g,
    y = compound,
    fill = source
  ) +
  scale_fill_manual(values = c("#3498eb", "black")) +
  coord_cartesian(expand = FALSE, xlim = c(0, 7800)) +
  geom_col() +
  theme_bw() +
  labs(
    x = "µmol / g"
  ) +
  theme(
    legend.position = "top",
    legend.title = element_blank(),
    axis.title.y = element_blank(),
    panel.grid = element_blank(),
    axis.text = element_text(color = "black")
  )
p_umol_g
```

Here we look at the fraction of exchangeable hydrogen for each biomolecule analyzed:

```{r}
p_frac_exch <- molecule_H_calc |> 
  mutate(frac_nHex = umol_Hex_per_g / umol_Htot_per_g) |> 
  mutate(compound = fct_reorder(.f = compound, .x = frac_nHex)) |> 
  ggplot() +
  aes(
    x = frac_nHex,
    y = compound,
    fill = group
  ) +
  geom_col(position = "dodge") +
  scale_x_continuous(labels = scales::percent) +
  coord_cartesian(xlim = c(0, 0.6), expand = FALSE) +
  theme_bw() +
  labs(
    x = "Fraction of Exchangeable H",
    y = "",
    fill = ""
  ) +
  theme(
    panel.grid = element_blank(),
    axis.text = element_text(color = "black")
  )
p_frac_exch

cowplot::save_plot(
  filename = "fig_output/frac_exch.pdf",
  plot = p_frac_exch,
  base_height = 6,
  base_width = 7
)
```

And here we examine the relative contributions of fractional H by each component of the cell:

```{r}
molecule_H_calc |> 
  mutate(compound = fct_reorder(.f = compound, .x = umol_Hex_per_g)) |> 
  ggplot() +
  aes(
    x = umol_Hex_per_g,
    y = compound,
    fill = group
  ) +
  geom_col(position = "dodge") +
  theme_bw() +
  labs(
    x = "µmol exchangeable H / g",
    y = "",
    fill = ""
  )
```

# Make a nice supplementary figure:

```{r}
p_combined_plot <- cowplot::plot_grid(
  p_mole_frac, p_frac_exch, p_umol_g,
  labels = c("A", "B", "C")
  )
p_combined_plot

# cowplot::save_plot(
#   filename = "fig_output/all_the_molecules.pdf",
#   plot = p_combined_plot,
#   base_height = 12,
#   base_width = 10
# )
```

```{r}
a <- tibble(
  b = c(0.79, 0.21),
  c = c("A", "B")
)

pa <- a |> ggplot() +
  aes(x = 1, y = b, fill = c) +
  geom_col() +
  theme_classic()

cowplot::save_plot(filename = "fig_output/bar.pdf",
                   plot = pa,
                   base_height = 8,
                   base_width = 5)
```
