==> 99_CD2F_DataViz.Rmd <==
---
title: "99_CD2F_DataViz"
author: "Tristan Caro"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---


# Setup

## Clear the environment

```{r}
rm(list=ls())
```


## Load libraries

```{r}
library(tidyverse) # CRAN v2.0.0
library(cowplot)   # CRAN v1.1.1
library(latex2exp) # CRAN v0.9.6
library(ggside)    # CRAN v0.2.2
library(ggsci)     # CRAN v3.0.0
library(ggdist)    # CRAN v3.3.0
library(ggpmisc)   # CRAN v0.5.4-1
library(ggpubr)    # CRAN v0.6.0
library(ggprism)   # CRAN v1.0.4
library(lans2r)
```

## Load from cache

```{r}
# Co-registered datasets
co_registered <- read_rds("cache/co_registered.RDS") # Raw
co_registered_fit <- read_rds("cache/co_registered_fitted.RDS") # Fitted
co_registered_summarized <- read_rds("cache/coreg_data_summary.rds") # sumamrized
  

# Dilution factor data
dilution_factors <- read_rds("cache/dilution_factors.rds")

# Area data
cd_pc_data <- read_rds("cache/cd_pc_data.RDS")
cd_pc_data_fitted <- read_rds("cache/area_data.RDS")

# Fitted xy data
fit_data <- read_rds("cache/fit_data.RDS")

# Fitting parameter data
fitted_xy_data <- read_rds("cache/peak_data_xy.RDS")

```

## Colors
```{r}
organism_colors <- c("#008c54", "#ffaa00")
```


# Plot


```{r linear model}
co_reg_fit <- broom::tidy(lm(data = co_registered, cd_pc_corr  ~ at2H_percent_dtc))

print_lm_summary <- function(df) {
  paste0("y = ", round(df$estimate[2], 2), "x + ", round(df$estimate[1], 2))
}

print_lm_summary(co_reg_fit)

co_reg_fit_thy <- broom::tidy(
  lm(data = co_registered %>%
       filter(
         organism.x == "T. hydrogeniphilus"
        ), 
     cd_pc_corr  ~ at2H_percent_dtc)
  )

co_reg_fit_mb <- broom::tidy(
  lm(data = co_registered %>%
       filter(
         organism.x == "M. NSHQ04"
        ), 
     cd_pc_corr  ~ at2H_percent_dtc)
  )

print_lm_summary(co_reg_fit_thy)
print_lm_summary(co_reg_fit_mb)
```



## Figure 1: CD / 2F ~ FL

### Fig. 1A: Example Spectra

Using cells "Thy_50_12_07", "Thy_00_07_23" as our examples.

```{r}
spectrum_palette = c("#0c008f", "#d43763")

p_spectra <- fitted_xy_data |> 
  filter(sample %in% c("Thy_00_07", "Thy_50_12")) |> 
  filter(cell_id %in% c("Thy_50_12_07", "Thy_00_07_23")) |> 
  mutate(
    Label = case_when(
      str_detect(cell_id, "Thy_50") ~ "50%",
      str_detect(cell_id, "Thy_00") ~ "0%",
    ),
    intensity = case_when(
      Label == "50%" ~ intensity + 100,
      Label == "0%" ~ intensity
    )
  ) |> 
  ggplot() +
  aes(
    x = wavenumber_cm,
    y = intensity,
    color = Label,
    fill = Label
  ) +
  geom_line(linewidth = 1, alpha = 0.8) +
  #facet_grid(vars(cell_id)) +
  scale_color_manual(values = spectrum_palette) +
  scale_fill_manual(values = spectrum_palette) +
  guides(x = "prism_offset_minor", y = "prism_offset_minor") + 
  labs(
    x = latex2exp::TeX("Wavenumber ($cm^{-1}$)"),
    y = "Intensity (AU)",
    color = "Growth Medium Deuterium\nEnrichment (at. %)",
    fill = "Growth Medium Deuterium\nEnrichment (at. %)"
  ) +
  theme_classic() +
  theme(
    legend.position = c(0.3, 0.8),
    legend.background = element_rect(color = "black"),
    strip.text = element_blank(),
    axis.text.x = element_text(color = "black", face = "bold"),
    axis.text.y = element_blank()
  )
p_spectra

cowplot::save_plot(
  plot = p_spectra,
  filename = "fig_output/Fig1A.pdf",
  base_height = 4,
  base_width = 6
)
```

### Fig. 1B: Example nanoSIMS

See python script


### Fig. 2 Summarized data
```{r}

# Summarized data:
co_registered_summarized_plot <- co_registered_summarized |> 
  select(-c(SE_cdpc, SE_2F)) |> 
  pivot_longer(
    cols = c(mean_2F, mean_cdpc),
    values_to = "value",
    names_to = "measurement"
  ) |> 
  mutate(
    sd = case_when(
      measurement == "mean_2F" ~ sd_2F,
      measurement == "mean_cdpc" ~ sd_cdpc
    )
  ) |> 
  mutate(
    measurement = case_when(
      measurement == "mean_2F" ~ "NanoSIMS",
      measurement == "mean_cdpc" ~ "Raman",
    )
  ) |> 
  mutate(
    # factor so that Raman appears first
    measurement = fct_relevel(measurement, "Raman")
  ) |>
  select(-c(sd_cdpc, sd_2F))


# Raw data for linear fit:
coreg_model_data_plot <- co_registered_fit |> 
  pivot_longer(
    cols = c(`CD%`, at2H_percent_dtc),
    values_to = "value",
    names_to = "measurement"
  ) |> 
  select(cell_id, organism, f_l, measurement, value) |> 
  mutate(
    measurement = case_when(
      measurement == "at2H_percent_dtc" ~ "NanoSIMS",
      measurement == "CD%" ~ "Raman",
    )) |> 
  mutate(
    # factor so that Raman appears first
    measurement = fct_relevel(measurement, "Raman")
  ) 
  
```

### Fig. 2 Output
```{r}
p_fig1 <- co_registered_summarized_plot |> 
  filter(f_l <= 40) |> 
  ggplot(
    aes(
      x = f_l,
      y = value
    )
  ) +
  # 1:1 line
  geom_abline(slope = 1, color = "black", linetype = "dotted") +
  # plot the lm from ALL the data, not the summarized points
  ggpmisc::stat_poly_line(
    data = coreg_model_data_plot |> filter(f_l != 50),
      aes(x = f_l,
          y = value),
      fullrange = TRUE,
      se = FALSE,
      size = 1,
      color = "#080099"
    ) +
  stat_poly_eq(
    data = coreg_model_data_plot |> filter(f_l != 50),
    use_label(c("eq", "R2"))
    ) +
  # the 50% point
  geom_pointinterval(
    size = 2,
    data = filter(co_registered_summarized_plot, f_l == 50),
    aes(x = f_l, y = value, ymin = value - sd, ymax = value + sd),
    color = "gray"
  ) +
  # the rest of the points
  geom_pointinterval(
    aes(
      fill = organism,
      ymin = value - sd,
      ymax = value + sd
    ),
    point_size = 2.7,
    shape = 21,
    color = "black"
  ) +
  scale_fill_manual(values = organism_colors) +
  facet_grid(measurement ~ organism) +
  scale_x_continuous(labels = scales::label_percent(scale = 1)) +
  scale_y_continuous(labels = scales::label_percent(scale = 1)) +
  ggprism::annotation_ticks(sides = "trbl") +
  coord_cartesian(xlim = c(0, 50), ylim = c(0,50)) +
  labs(
    x = latex2exp::TeX("Growth Water Deuterium Enrichment ($^2F_{gw}$)"),
    y = latex2exp::TeX("Biomass Deuterium Enrichment ($^2F_{Biomass}$)")
  ) +
  theme_classic() +
  theme(
    legend.position = "NA",
    strip.background = element_blank(),
    panel.border = element_rect(color = "black", fill = "NA", size = 1),
    axis.line = element_blank(),
    strip.background.x = element_rect(fill = "black", size = 1),
    strip.background.y = element_rect(fill = "#363636", color = "#363636", size = 1),
    strip.text = element_text(color = "white", size = 12, face = "bold"),
    strip.text.x = element_text(face = "italic"),
    aspect.ratio = 1,
    axis.text = element_text(face = "bold", color = "black"),
    axis.title = element_text(face = "bold", color = "black")
  )

p_fig1

cowplot::save_plot(
  plot = p_fig1, 
  filename = "fig_output/Fig2.pdf",                     
  base_height = 6, base_width = 6
  )
```




## CD + 2Fbc ~ FL

This plot represents 2F calculated by CD%, 2F calculated by back-calculating true nanoSIMS values using dilution factors (DF), plotted against label strength (FL)

```{r}
co_registered_summarized_mini <- co_registered_summarized |> 
  select(f_l, mean_cdpc, sd_cdpc, n)

dilution_factors_mini <- dilution_factors |> 
  select(f_l, F_before_backcalc, sF_before_backcalc)

co_registered_summarized_mini |> 
  left_join(dilution_factors_mini, by = "f_l") |> 
  filter(f_l != 0) |>
  pivot_longer(cols = c(mean_cdpc, F_before_backcalc), 
            names_to = "method", 
            values_to = "2F"
            ) |> 
  pivot_longer(
    cols = c(sd_cdpc, sF_before_backcalc),
    names_to = "error_type",
    values_to = "SD"
  ) |> 
  ggplot(
    aes(x = f_l,
        y = `2F`,
        color = method)
  ) +
  geom_point() +
  geom_pointrange(
    aes(ymin = `2F` - SD,
        ymax = `2F` + SD,
        color = error_type)
  ) +
  facet_wrap(vars(error_type)) +
  theme_bw()
  
```

## Supp: Co-registered

```{r co-registered_faceted}
p_co_registered <- co_registered %>% 
  filter(!is.na(f_l)) %>% 
  # Plot it!
  ggplot(
    aes(y = at2H_percent_dtc,
        x = cd_pc_corr,
        #color = as.factor(f_l),
        #group = organism.x,
        #label = cell_id
    )
  ) +
   ggpmisc::stat_poly_line(
                          fullrange = TRUE,
                          se = FALSE,
                          size = 0.3,
                          color = "#363636") +
  stat_poly_eq(use_label(c("eq", "R2"))) +
  geom_point(
    alpha = 0.7,
    size = 1.5
  ) +
  geom_abline(
    slope = 1,
    linetype = "dotted",
    color = "gray",
    alpha = 1,
  ) +
  #facet_wrap(vars(organism.x)) +
  coord_cartesian(
    xlim = c(0,40), 
    ylim = c(0,40)
  ) +
  scale_color_viridis_d(
    end = 0.9,
    option = "magma"
    ) +
  theme_bw() +
  labs(
    y = latex2exp::TeX("$^2F\\%_{nanoSIMS}$"),
    x = latex2exp::TeX("$^2F\\%_{Raman}$"),
    color = latex2exp::TeX("$^2H_2O$ Label Strength $^2F\\%$"),
    shape = ""
  ) +
  theme(
    legend.position = "bottom",
    aspect.ratio = 1
  )
p_co_registered

cowplot::save_plot(
  plot = p_co_registered,
  filename = "fig_output/figS1.pdf",
  base_height = 3,
  base_width = 3
)

cowplot::save_plot(
  plot = p_co_registered,
  filename = "fig_output/figS1.png",
  base_height = 3,
  base_width = 3
)
```


## Supp: Deadtime Correction Effect

```{r}
co_registered_lm <- lm(data = co_registered, at2H_percent_dtc ~ at2H_percent) |> 
  broom::tidy()


co_registered_dtc <- co_registered |> 
  select(at2H_percent_dtc, at2H_percent, f_l) |> 
  filter(!is.na(f_l)) |> 
  ungroup() |> 
  mutate(
    at2H_residual= at2H_percent - at2H_percent_dtc) |>
  mutate(n = as.character(row_number())) |> 
  mutate(n = forcats::fct_reorder(n, -at2H_residual)) |> 
  arrange(-at2H_residual) |> 
  mutate(n_n = row_number())

p_dtc_A <- co_registered_dtc |> ggplot() +
  aes(
    x = n_n,
    y = at2H_residual,
    color = as.factor(f_l)
  ) +
  geom_point(alpha = 1, size = 1) +
  ggside::geom_xsidedensity() +
  scale_color_viridis_d(option = "plasma", begin = "0", end = 0.8) +
  theme_bw() +
  labs(
    y = latex2exp::TeX("($^2F_{raw}$ - $^2F_{DTC}$) %"),
    x = "",
    color = latex2exp::TeX("$^2F_{label}$ %"),
  ) +
  theme(
    legend.position = "none",
    ggside.panel.scale.x = .4,
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank()
  )


p_dtc_B <- co_registered_dtc |> ggplot() +
  aes(
    x = at2H_percent,
    y = at2H_residual,
    color = as.factor(f_l)
  ) +
  geom_point() +
  geom_smooth(method = "lm", formula = "y~x", color = "red", size = 0.5, se = FALSE) +
  ggpubr::stat_regline_equation(
     aes(label =  paste(..eq.label.., ..rr.label.., sep = "~~~~")),
     color = "red",
     size = 2
  ) +
  scale_color_viridis_d(option = "plasma", begin = "0", end = 0.8) +
  labs(
    y = latex2exp::TeX("($^2F_{raw}$ - $^2F_{DTC}$) %"),
    x = latex2exp::TeX("$^2F_{raw}$ %"),
    color = latex2exp::TeX("$^2F_{label}$ %"),
  ) +
  theme_bw() +
  theme(
    legend.position = "none"
  )
p_dtc_B


p_dtc_C <- co_registered_dtc |> ggplot() +
  aes(
    x = at2H_percent,
    y = at2H_percent_dtc
  ) +
  geom_point(aes(color = as.factor(f_l))) +
  geom_smooth(method = "lm", formula = "y~x", color = "red", size = 0.5) +
  ggpubr::stat_regline_equation(
     aes(label =  paste(..eq.label.., ..rr.label.., sep = "~~~~")),
     color = "red"
  ) +
  scale_color_viridis_d(option = "plasma", begin = "0", end = 0.8) +
  labs(
    x = latex2exp::TeX("$^2F_{raw}$%"),
    y = latex2exp::TeX("$^2F_{DTC}$%"),
    color = latex2exp::TeX("$^2F_{label}$ %")
  ) +
  theme_bw() +
  theme(
    legend.position = "bottom"
  )
p_dtc_C

top_row <- cowplot::plot_grid(p_dtc_A, p_dtc_B, labels = c("A", "B"), rel_widths = c(0.5, 0.5))
p_dtc <- cowplot::plot_grid(top_row, p_dtc_C, nrow = 2, labels = c(" ", "C"), rel_heights = c(0.4, 0.6))

cowplot::save_plot(plot = p_dtc, filename = "fig_output/dtc_effect.pdf",
                   base_height = 6, base_width = 6)
cowplot::save_plot(plot = p_dtc, filename = "fig_output/dtc_effect.png",
                   base_height = 6, base_width = 6)

```


## Supp: T. hy vs. NSHQ04 spectra

```{r}
p_species_spectrum <- cd_pc_data |> 
  filter(cell_id %in% c("Thy_50_12_25", "Mb_50_18_32")) |> 
  mutate(
    organism = case_when(
      str_detect(cell_id, "Thy") ~ "T. hydrogeniphilus",
      str_detect(cell_id, "Mb") ~ "M. NSHQ04"
    )
  ) |> 
  ggplot() +
  aes(
    x = wavenumber_cm,
    y = intensity,
    color = organism
  ) +
  geom_line() +
  facet_wrap(vars(cell_id), scales = "free_y", ncol = 1) +
  scale_color_manual(values = organism_colors) +
  scale_fill_manual(values = organism_colors) +
  labs(
    x = latex2exp::TeX("Wavenumber ($cm^{-1}$)"),
    y = "Intensity (AU)",
    color = "",
    fill = ""
  ) +
  theme_classic() +
  theme(
    legend.position = "top",
    legend.background = element_rect(color = "black"),
    strip.text = element_blank(),
    axis.text.x = element_text(color = "black", face = "bold"),
    axis.text.y = element_blank()
  )

cowplot::save_plot(
  filename = "fig_output/species_spectrum.pdf",
  p_species_spectrum,
  base_height = 8,
  base_width = 8
)
```


## Raman spectrum

```{r}
cd_pc_data %>% filter(cell_id == "Mb_50_18_38") %>% 
  ggplot(aes(
    x = wavenumber_cm,
    y = intensity
  )) +
  geom_line() +
  theme_minimal()
```


## Compare fitted co-reg with raw data

### Compare the co-registered datasets

```{r}
co_reg_compare <- bind_rows(
  co_registered %>% select(at2H_percent_dtc, cd_pc, f_l, organism.x) %>% mutate(type = "raw", organism = organism.x),
  co_registered_fit %>% select(at2H_percent_dtc, `CD%`, f_l, organism) %>% mutate(type = "fitted")
) %>%
  mutate(cd_percent = case_when(
    is.na(cd_pc) ~ `CD%`,
    is.na(`CD%`) ~ cd_pc
  )) %>% 
  select(-c(cd_pc, `CD%`, organism.x))
```

```{r}
co_reg_compare %>% 
  ggplot(
    aes(
      x = at2H_percent_dtc,
      y = cd_percent,
      color = type
    )
  ) +
  geom_point(
    alpha = 0.5
  ) +
  geom_smooth(
    method = "lm",
    formula = "y ~ x"
  ) +
  geom_abline(
    slope = 1,
    linetype = "dashed",
    color = "gray",
    alpha = 1,
  ) +
  facet_wrap(vars(type)) +
  coord_cartesian(
    xlim = c(0,50), 
    ylim = c(0,50)
  ) +
  scale_color_viridis_d(
    end = 0.7,
    begin = 0.3,
    option = "magma"
    ) +
  theme_bw() +
  labs(
    x = latex2exp::TeX("$^2F\\%_{nanoSIMS}$"),
    y = latex2exp::TeX("$^2F\\%_{Raman}$"),
    color = latex2exp::TeX("$^2H_2O$ Label Strength $^2F\\%$")
  ) +
  theme(
    legend.position = "bottom"
  )
```

### CD_FL jitter plot

```{r}
p_cd_fl1 <- co_registered %>% 
  filter(!is.na(f_l)) %>% 
  ggplot(
    aes(x = f_l,
        y = cd_pc_corr,
        color = organism.x
        )
  ) +
  labs(
    x = latex2exp::TeX("$^2F$ Growth Water"),
    y = latex2exp::TeX("$^2F_{Raman}$"),
    title = "Raw Data"
  ) +
  facet_wrap(
    vars(organism.x), 
    scales = "free") +
  geom_abline(
    slope = 1,
    linetype = "dashed",
    color = "gray",
    alpha = 1,
  ) +
  geom_point(
    alpha = 0.5,
    position = position_jitter(width = 2)
  ) +
  ggdist::stat_pointinterval(
    position = position_nudge(x = 0.2),
    color = "black"
  ) +
  scale_color_manual(values = organism_colors) + 
  scale_x_continuous(
    limits = c(0, 55),
    breaks = c(0, 10, 20, 30, 40, 50)
    ) +
  scale_y_continuous(
    limits = c(0, 55),
    breaks = c(0, 10, 20, 30, 40, 50)
  ) +
  theme_classic() +
  theme(
    legend.position = "NA",
    strip.background = element_blank()
  )

p_cd_fl2 <- co_registered_fit %>% 
  filter(!is.na(f_l)) %>% 
  ggplot(
    aes(x = f_l,
        y = `CD%`,
        color = organism
        )
  ) +
  labs(
    x = latex2exp::TeX("$^2F$ Growth Water"),
    y = latex2exp::TeX("$^2F_{Raman}$"),
    title = "Fitted Data"
  ) +
  facet_wrap(
    vars(organism), 
    scales = "free") +
  geom_abline(
    slope = 1,
    linetype = "dashed",
    color = "gray",
    alpha = 1,
  ) +
  geom_point(
    alpha = 0.5,
    position = position_jitter(width = 2),
  ) +
  ggdist::stat_pointinterval(
    position = position_nudge(x = 0.2),
    color = "black"
  ) +
  scale_color_manual(values = organism_colors) + 
  scale_x_continuous(
    limits = c(0, 55),
    breaks = c(0, 10, 20, 30, 40, 50)
    ) +
  scale_y_continuous(
    limits = c(0, 55),
    breaks = c(0, 10, 20, 30, 40, 50)
  ) +
  theme_classic() +
  theme(
    legend.position = "NA",
    strip.background = element_blank()
  )

co_reg_compared <- cowplot::plot_grid(
  p_cd_fl1,
  p_cd_fl2,
  ncol = 1
)
co_reg_compared

save_plot(
  filename = "fig_output/co_reg_compare.pdf",
  plot = co_reg_compared,
  base_height = 8,
  base_width = 8
  )


```

### CD_FL boxplot comparison

```{r}
co_reg_compare %>% 
  filter(!is.na(f_l)) %>% 
  ggplot(
    aes(
      x = as.factor(f_l),
      y = cd_percent,
      color = type
    )
  ) +
  ggdist::stat_dist_pointinterval(position = position_dodgejust(width = 0.5)) +
  scale_color_manual(values = c("black", "red")) +
  labs(
    x = latex2exp::TeX("$^2H_2O$ Label Strength $^2F\\%$"),
    y = latex2exp::TeX("$^2F\\%_{Raman}$"),
    color = "Data Source"
  ) +
  theme_classic()
```


# Fitting parameter data
```{r}
p_fit_thy <- fit_data %>% 
  mutate(organism =
    case_when(
      str_detect(sample, "Thy") ~ "T. hydrogeniphilus",
      str_detect(sample, "Mb") ~ "M. NSHQ04"
      )
  ) %>% 
  pivot_longer(
    cols = -c(fit, sample, organism),
    names_to = "fit_param",
    values_to = "value"
  ) %>%
  filter(organism == "T. hydrogeniphilus") %>% 
  ggplot(
    aes(x = value,
        fill = organism)
  ) +
  geom_histogram(
    bins = 100,
    alpha = 0.5,
    fill = "orange"
    ) +
  facet_wrap(vars(fit_param, fit), scales = "free") +
  theme_bw()

p_fit_mb <- fit_data %>% 
  mutate(organism =
    case_when(
      str_detect(sample, "Thy") ~ "T. hydrogeniphilus",
      str_detect(sample, "Mb") ~ "M. NSHQ04"
      )
  ) %>% 
  pivot_longer(
    cols = -c(fit, sample, organism),
    names_to = "fit_param",
    values_to = "value"
  ) %>%
  filter(organism == "M. NSHQ04") %>% 
  ggplot(
    aes(x = value,
        fill = organism)
  ) +
  geom_histogram(
    bins = 100,
    alpha = 0.5,
    fill = "green"
    ) +
  facet_wrap(vars(fit_param,fit ), scales = "free") +
  theme_bw()


p_fit_mb

cowplot::plot_grid(p_fit_thy, p_fit_mb)
```
